# Port Hibernate's EntityKey optimization