@fredericDelaporte 2019-01-13
For merging, I intend to squash and merge in order to reword the commit.