@hazzik 2019-01-13
Need to implement IEquitable to squeeze more performance.