@maca88 2019-01-13
The results are more or less the same: 
```
Async eager Load fetches
-------------------------
# of elements fetched: 6768 (4768 + 1000 + 1000).    Fetch took: 87.025.824,00ms. Allocated bytes: 87025824.

Results per framework. Values are given as: 'mean (standard deviation)'
==============================================================================

Change tracking fetches, set fetches (20 runs), no caching
------------------------------------------------------------------------------
NHibernate v5.2.0.0 (v5.2.2.0) : 1.851,49ms (51,74ms)   Enum: 1,72ms (0,11ms)

Memory usage, per iteration
------------------------------------------------------------------------------
NHibernate v5.2.0.0 (v5.2.2.0) : 173.071 KB (177.225.576 bytes)
```