@bahusoid 2019-01-14
@maca88  I've prepared a patch for this branch with EntityKey class->struct conversion:
[0001-Convert-EntityKey-to-struct.zip](https://github.com/nhibernate/nhibernate-core/files/2754874/0001-Convert-EntityKey-to-struct.zip)
It can be applied via `git am < 0001-Convert-EntityKey-to-struct.patch` for unzipped file. Could you please check how it affects benchmark results.