@bahusoid 2019-01-14
Interesting. Thanks! Though I'm not sure I'm following you regarding diffreneces for `string` vs `int` for identifier. From what I know stored value in field should not affect struct size in any way.