@maca88 2019-01-14
As string is not a value type, a reference to the string will be stored in the struct which on x64 platform is 8 bytes. So if we combine all fields we get 20 bytes:
identifier (string): 8 bytes (reference)
persister: 8 bytes (reference)
hashcode: 4 byte (value)