@bahusoid 2019-01-14
What I'm saying is `EntityKey.identifier` is of type `object`. I expect it always to be 8 bytes on 64 bit. It doesn't matter what value you put in it -  boxed int or string