@hazzik 2019-01-14
>By just appling your patch the overall performance wasn't changed

But memory usage dropped by ~6 megabytes.