@fredericDelaporte 2019-01-14
> Another option for optimizing such cases is to use the object pooling that is frequently [used in games](https://www.raywenderlich.com/847-object-pooling-in-unity).

This pooling implies mutating the cache keys. I find it quite debatable: it has a good potential for serious bugs. If a key, which has been used for storing caching something, is mutated, things will obviously go wrong.