@maca88 2019-01-15
> This pooling implies mutating the cache keys. I find it quite debatable: it has a good potential for serious bugs

Indeed, it may be the fastest but for this case using a struct would probably be a better choice which would also lower the memory usage as @hazzik pointed out.